move source files into "strato_certbot" subfolder, renamed script files to avoid hyphens, added __init__.py to indicate package
